refactor(workspace): 패널 리사이즈 상태를 전용 훅으로 분리
EditorWorkspace 안에 몰려 있던 패널 폭 상태, pointer/keyboard 리사이즈 처리, 커밋 타이밍을 usePanelResizing으로 이동했다.

이 변경으로 EditorWorkspace는 패널을 어떻게 렌더링하는지에 집중하고, 리사이즈 상태 전이와 폭 계산 규칙은 별도 훅에서 관리한다.

추가로 usePanelResizing 테스트를 도입해 키보드 리사이즈와 pointer 리사이즈 커밋 경로를 검증했다.

검증:
- npm run test
