+++ src/components/workspace/EditorWorkspace.tsx
@@ -1,4 +1,4 @@
-import type { KeyboardEvent as ReactKeyboardEvent, PointerEvent as ReactPointerEvent, RefObject } from "react";
+import type { RefObject } from "react";
 import { useDeferredValue, useEffect, useEffectEvent, useMemo, useRef, useState } from "react";
 import type { MarkdownEditorHandle } from "../editor/MarkdownEditor";
 import { MarkdownEditor } from "../editor/MarkdownEditor";
@@ -18,9 +18,9 @@ import { useDebouncedValue } from "../../hooks/useDebouncedValue";
 import { useIdleValue } from "../../hooks/useIdleValue";
 import {
   clampPanelWidth,
-  getEffectivePanelWidths,
   getMaxAllowedPanelWidth,
 } from "./panel-layout";
+import { usePanelResizing } from "./usePanelResizing";
 
 const PREVIEW_DEBOUNCE_MS = 120;
 const PREVIEW_IDLE_TIMEOUT_MS = 250;
@@ -45,13 +45,7 @@ type EditorWorkspaceProps = {
   onEditorFocusChange: (focused: boolean) => void;
 };
 
-type PanelKind = "preview" | "toc";
 type PanelVisibilityState = "closed" | "entering" | "closing" | "open";
-
-type PanelWidthsState = {
-  preview: number | null;
-  toc: number | null;
-};
 
 function getReducedMotionPreference() {
   if (typeof window === "undefined" || typeof window.matchMedia !== "function") {
@@ -274,23 +268,7 @@ export function EditorWorkspace({
   onEditorFocusChange,
 }: EditorWorkspaceProps) {
   const mainRef = useRef<HTMLElement | null>(null);
-  const dragStateRef = useRef<{
-    initialWidth: number;
-    kind: PanelKind;
-    pointerId: number;
-    startX: number;
-  } | null>(null);
-  const latestPanelWidthsRef = useRef<PanelWidthsState>({
-    preview: initialPreviewPanelWidth,
-    toc: initialTocPanelWidth,
-  });
   const [containerWidth, setContainerWidth] = useState<number | null>(null);
-  const [layoutVersion, setLayoutVersion] = useState(0);
-  const [activeResizer, setActiveResizer] = useState<PanelKind | null>(null);
-  const [panelWidths, setPanelWidths] = useState<PanelWidthsState>({
-    preview: initialPreviewPanelWidth,
-    toc: initialTocPanelWidth,
-  });
   const previewPresence = usePanelPresence(isPreviewVisible);
   const tocPresence = usePanelPresence(isTocVisible);
   const markdown = useDeferredValue(useDocumentMarkdown(documentStore));
@@ -311,34 +289,21 @@ export function EditorWorkspace({
     previewPresence.state === "closing" ||
     tocPresence.state === "entering" ||
     tocPresence.state === "closing";
-  const isResizingPanels = activeResizer !== null;
-
-  useEffect(() => {
-    setPanelWidths({
-      preview: initialPreviewPanelWidth,
-      toc: initialTocPanelWidth,
-    });
-    latestPanelWidthsRef.current = {
-      preview: initialPreviewPanelWidth,
-      toc: initialTocPanelWidth,
-    };
-  }, [initialPreviewPanelWidth, initialTocPanelWidth]);
-
-  const effectivePanelWidths = useMemo(() => {
-    return getEffectivePanelWidths({
-      containerWidth,
-      hasPreview: hasRenderedPreview,
-      hasToc: hasRenderedToc,
-      previewWidth: panelWidths.preview,
-      tocWidth: panelWidths.toc,
-    });
-  }, [
+  const {
+    activeResizer,
+    effectivePanelWidths,
+    isResizingPanels,
+    layoutVersion,
+    resizeWithKeyboard,
+    startResize,
+  } = usePanelResizing({
     containerWidth,
-    hasRenderedPreview,
-    hasRenderedToc,
-    panelWidths.preview,
-    panelWidths.toc,
-  ]);
+    hasPreview: hasRenderedPreview,
+    hasToc: hasRenderedToc,
+    initialPreviewPanelWidth,
+    initialTocPanelWidth,
+    onPanelWidthsChange,
+  });
 
   useEffect(() => {
     const container = mainRef.current;
@@ -381,179 +346,6 @@ export function EditorWorkspace({
       onPathCopyError();
     }
   });
-
-  const commitPanelWidths = useEffectEvent((nextPanelWidths: PanelWidthsState) => {
-    onPanelWidthsChange({
-      previewPanelWidth: nextPanelWidths.preview,
-      tocPanelWidth: nextPanelWidths.toc,
-    });
-  });
-
-  const applyPanelWidth = useEffectEvent((kind: PanelKind, nextWidth: number) => {
-    setPanelWidths((current) => {
-      const nextPanelWidths = kind === "toc"
-        ? {
-            ...current,
-            toc: nextWidth,
-          }
-        : {
-            ...current,
-            preview: nextWidth,
-          };
-
-      latestPanelWidthsRef.current = nextPanelWidths;
-
-      return nextPanelWidths;
-    });
-  });
-
-  const updatePanelWidths = useEffectEvent((nextPanelWidths: PanelWidthsState) => {
-    latestPanelWidthsRef.current = nextPanelWidths;
-    setPanelWidths(nextPanelWidths);
-  });
-
-  const finalizeResize = useEffectEvent(() => {
-    dragStateRef.current = null;
-    setActiveResizer(null);
-    setLayoutVersion((version) => version + 1);
-    commitPanelWidths(latestPanelWidthsRef.current);
-  });
-
-  useEffect(() => {
-    latestPanelWidthsRef.current = panelWidths;
-  }, [panelWidths]);
-
-  useEffect(() => {
-    const handlePointerMove = (event: PointerEvent) => {
-      const dragState = dragStateRef.current;
-      if (!dragState) {
-        return;
-      }
-
-      const deltaX = event.clientX - dragState.startX;
-      const requestedWidth = dragState.kind === "toc"
-        ? dragState.initialWidth + deltaX
-        : dragState.initialWidth - deltaX;
-      const nextWidth = clampPanelWidth(dragState.kind, requestedWidth, {
-        containerWidth,
-        hasPreview: hasRenderedPreview,
-        hasToc: hasRenderedToc,
-        siblingWidth: dragState.kind === "toc"
-          ? effectivePanelWidths.previewWidth
-          : effectivePanelWidths.tocWidth,
-      });
-
-      applyPanelWidth(dragState.kind, nextWidth);
-    };
-
-    const handlePointerUp = (event: PointerEvent) => {
-      const dragState = dragStateRef.current;
-      if (!dragState || dragState.pointerId !== event.pointerId) {
-        return;
-      }
-
-      finalizeResize();
-    };
-
-    window.addEventListener("pointermove", handlePointerMove);
-    window.addEventListener("pointerup", handlePointerUp);
-    window.addEventListener("pointercancel", handlePointerUp);
-
-    return () => {
-      window.removeEventListener("pointermove", handlePointerMove);
-      window.removeEventListener("pointerup", handlePointerUp);
-      window.removeEventListener("pointercancel", handlePointerUp);
-    };
-  }, [
-    applyPanelWidth,
-    containerWidth,
-    effectivePanelWidths.previewWidth,
-    effectivePanelWidths.tocWidth,
-    finalizeResize,
-    hasRenderedPreview,
-    hasRenderedToc,
-  ]);
-
-  function startResize(kind: PanelKind, event: ReactPointerEvent<HTMLDivElement>) {
-    const initialWidth = kind === "toc"
-      ? effectivePanelWidths.tocWidth
-      : effectivePanelWidths.previewWidth;
-    if (initialWidth === null) {
-      return;
-    }
-
-    dragStateRef.current = {
-      initialWidth,
-      kind,
-      pointerId: event.pointerId,
-      startX: event.clientX,
-    };
-    setActiveResizer(kind);
-    event.preventDefault();
-  }
-
-  function resizeWithKeyboard(kind: PanelKind, event: ReactKeyboardEvent<HTMLDivElement>) {
-    const step = event.shiftKey ? 48 : 16;
-    const currentWidth = kind === "toc"
-      ? effectivePanelWidths.tocWidth
-      : effectivePanelWidths.previewWidth;
-
-    if (currentWidth === null) {
-      return;
-    }
-
-    const siblingWidth = kind === "toc"
-      ? effectivePanelWidths.previewWidth
-      : effectivePanelWidths.tocWidth;
-    const maxWidth = getMaxAllowedPanelWidth({
-      containerWidth,
-      hasPreview: hasRenderedPreview,
-      hasToc: hasRenderedToc,
-      kind,
-      siblingWidth,
-    });
-
-    let nextWidth: number | null = null;
-
-    if (event.key === "Home") {
-      nextWidth = clampPanelWidth(kind, 0, {
-        containerWidth,
-        hasPreview: hasRenderedPreview,
-        hasToc: hasRenderedToc,
-        siblingWidth,
-      });
-    } else if (event.key === "End") {
-      nextWidth = maxWidth;
-    } else if (kind === "toc" && event.key === "ArrowLeft") {
-      nextWidth = currentWidth - step;
-    } else if (kind === "toc" && event.key === "ArrowRight") {
-      nextWidth = currentWidth + step;
-    } else if (kind === "preview" && event.key === "ArrowLeft") {
-      nextWidth = currentWidth + step;
-    } else if (kind === "preview" && event.key === "ArrowRight") {
-      nextWidth = currentWidth - step;
-    }
-
-    if (nextWidth === null) {
-      return;
-    }
-
-    event.preventDefault();
-
-    const clampedWidth = clampPanelWidth(kind, nextWidth, {
-      containerWidth,
-      hasPreview: hasRenderedPreview,
-      hasToc: hasRenderedToc,
-      siblingWidth,
-    });
-    const nextPanelWidths = kind === "toc"
-      ? { ...panelWidths, toc: clampedWidth }
-      : { ...panelWidths, preview: clampedWidth };
-
-    updatePanelWidths(nextPanelWidths);
-    setLayoutVersion((version) => version + 1);
-    commitPanelWidths(nextPanelWidths);
-  }
 
   return (
     <EditorViewStateProvider documentKey={documentKey}>
